Trait boogie_backend::prover_task_runner::ProverTask

pub trait ProverTask {
    type TaskResult: Send + 'static;
    type TaskId: Send + Copy + 'static;
    fn init(&mut self, num_instances: usize) -> Vec<Self::TaskId>;

#[must_use]    fn run<'life0, 'async_trait>(
        &'life0 mut self, 
        task_id: Self::TaskId, 
        sem: Arc<Semaphore>
    ) -> Pin<Box<dyn Future<Output = Self::TaskResult> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn is_success(&self, task_result: &Self::TaskResult) -> bool;
}

Associated Types

type TaskResult: Send + 'static

type TaskId: Send + Copy + 'static

Loading content...
Required methods

fn init(&mut self, num_instances: usize) -> Vec<Self::TaskId>

Initialize the task runner given the number of instances.

#[must_use]fn run<'life0, 'async_trait>(
&'life0 mut self,
task_id: Self::TaskId,
sem: Arc<Semaphore>
) -> Pin<Box<dyn Future<Output = Self::TaskResult> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Run the task with task_id. This function will be called from one of the worker threads.

fn is_success(&self, task_result: &Self::TaskResult) -> bool

Returns whether the task result is considered successful.

Loading content...